% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER.get_top_pathway_LVs}
\alias{autoPLIER.get_top_pathway_LVs}
\title{AutoPLIER Get Top Pathways LVs Method}
\usage{
autoPLIER.get_top_pathway_LVs(ap, ...)
}
\arguments{
\item{pathway}{List of pathways.}

\item{n_LVs}{Number of latent variables to retrieve.}
}
\description{
AutoPLIER Get Top Pathways LVs Method
}
