% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER.get_top_pathways}
\alias{autoPLIER.get_top_pathways}
\title{AutoPLIER Get Top Pathways Method}
\usage{
autoPLIER.get_top_pathways(ap, ...)
}
\arguments{
\item{LVs}{List of decomposed latent variables.}

\item{n_pathways}{Number of pathways to retrieve.}
}
\description{
AutoPLIER Get Top Pathways Method
}
