% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER}
\alias{autoPLIER}
\title{AutoPLIER class}
\usage{
autoPLIER(...)
}
\arguments{
\item{n_components}{Number of latent variables/latent dimensions.}

\item{regval}{Regularization value. Default: \code{1.20E-7}.}

\item{learning_rate}{Proportion that weights are updated during training.
Default: \code{0.0002}.}

\item{scaler}{Scaler to transform data.}
}
\value{
A Python object called \code{autoplier.model.autoPLIER} that cannot be
viewed.
}
\description{
Calls the constructor autoPLIER to create a model.
}
