% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER.fit_transform}
\alias{autoPLIER.fit_transform}
\title{AutoPLIER Fit-Transform Method}
\usage{
autoPLIER.fit_transform(ap, ...)
}
\arguments{
\item{x_train}{Training dataset used to train the model.}

\item{pathways}{Pathways dataset.}

\item{callbacks}{List of callbacks to apply during training.}

\item{batch_size}{Fixed batch size.}

\item{maxepoch}{Max number of epochs.}

\item{verbose}{Verbosity mode. 0 = silent, 1 = progress bar,
2 = one line per epoch.}

\item{valfrac}{Fraction of the training dataset used as validation data.}
}
\description{
AutoPLIER Fit-Transform Method
}
