% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER.fit_transform}
\alias{autoPLIER.fit_transform}
\title{AutoPLIER Fit-Transform Method}
\usage{
autoPLIER.fit_transform(ap, mod, ...)
}
\arguments{
\item{ap}{AutoPLIER object.}

\item{x_train}{Training dataset used to train the model.}

\item{pathways}{Pathways dataset.}

\item{callbacks}{List of callbacks to apply during training.}

\item{batch_size}{Fixed batch size. Default: \code{NULL}.}

\item{maxepoch}{Max number of epochs. Default: \code{2000}.}

\item{verbose}{Verbosity mode. 0 = silent, 1 = progress bar,
2 = one line per epoch. Default: \code{2}.}

\item{valfrac}{Fraction of the training dataset used as validation data.
Default: \code{0.3}.}
}
\value{
A dataframe consisting of a number of LV columns as specified by the
\code{n_components} in \code{autoPLIER()}.
}
\description{
Simultaneously fit the data and transform the autoPLIER model.
}
