% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER.get_top_pathway_LVs}
\alias{autoPLIER.get_top_pathway_LVs}
\title{AutoPLIER Get Top Pathways LVs Method}
\usage{
autoPLIER.get_top_pathway_LVs(ap, ...)
}
\arguments{
\item{ap}{AutoPLIER object.}

\item{pathway}{List of pathways.}

\item{n_LVs}{Number of latent variables to retrieve.}
}
\value{
Numeric object with the chosen pathway and its corresponding LVs.
}
\description{
Retrieve LVs associated with a named pathway.
}
