% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER}
\alias{autoPLIER}
\title{AutoPLIER class}
\usage{
autoPLIER(...)
}
\arguments{
\item{n_components}{Number of latent variables/latent dimensions.}

\item{regval}{Regularization value.}

\item{learning_rate}{Proportion that weights are updated during training.}

\item{scaler}{Scaler to transform data.}
}
\description{
AutoPLIER class
}
