% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER.transform}
\alias{autoPLIER.transform}
\title{AutoPLIER Transform Method}
\usage{
autoPLIER.transform(ap, ...)
}
\arguments{
\item{x_predict}{Dataset to be transformed into the latent space.}

\item{pathways}{Pathways dataset.}
}
\description{
AutoPLIER Transform Method
}
