% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoPLIER.R
\name{autoPLIER.transform}
\alias{autoPLIER.transform}
\title{AutoPLIER Transform Method}
\usage{
autoPLIER.transform(ap, ...)
}
\arguments{
\item{ap}{AutoPLIER object.}

\item{x_predict}{Dataset to be transformed into the latent space.}

\item{pathways}{Pathways dataset.}
}
\value{
A dataframe consisting of a number of LV columns as specified by the
\code{n_components} in \code{autoPLIER()}.
}
\description{
Applies the fitted calculation to the created autoPLIER model.
}
